Expand Capybara wait time blocks to reduce test flakiness
What?
=====

This test performs a combination of assertions against Capybara's max
wait time in conjunction with current Capybara behavior around when to
wait for an element to be found on the page to verify at an integration
level that we're using the correct underlying Capybara behavior for
assertions.

However, by using this, and based on timing and hardware on CI, this
inherently has test jitter, resulting in periodic failures.

This expands the time windows slightly to be slightly looser in its
assertions while still building confidence we're using the correct
functionality from Capybara.