Extract method delegation to separate module
What?
=====

This introduces a new module, `DelegatesTo`, which allows for dynamic
definitions of both `method_missing` and `respond_to_missing?` based on
the argument to the included module.

Available delegation/decoration mechanisms in Ruby don't handle the use
case; for example, `Delegator`/`SimpleDelegator` rely on defining the
wrapped object via `__getobj__`, while `DelegateClass` require
specifying the class to be decorated. `Forwardable` and Rails'
`delegate` require explicit declaration of all methods to be delegated,
which isn't realistic for `Capybara::Node::Element`s given the object's
method surface area.

This uses metaprogramming to generate a module at runtime that is mixed
into the included class, allowing the caller to specify the delegation
target.