Raise when declared elements conflict with Capybara's RSpec matchers
What?
=====

PageEz defines predicate methods for each element declared, making it
easy to assert against presence or absence on a page. RSpec
then takes these predicate methods to generate implicit matchers.

However, concrete matchers take precedence over predicate matchers. In
the case of elements that may feasibly be defined in a page object, like
`title`, `button`, `field`, `select`, and `table`, this may present
challenges if arguments are passed that differ from the matcher
implementation.

This introduces a new configuration setting, which defaults to raising
an exception, when a collision occurs.