Adjust DelegatesTo to mix in a module rather than define methods directly
What?
=====

The previous version of DelegatesTo generated an anonymous module that,
when included, would define methods directly on the object that included
it.

This commit adjusts the behavior slightly to include a new module with
method_missing and respond_to_missing? defined, allowing for the object
including the module to call super to bubble up delegation behavior.