Refactor to method definition strategies
What?
=====

This adjusts the previous approach to defining methods by extracting
method definition strategies.

By continuing to use the pattern matching on args within the has_one,
this constructs different strategies based on known values, which then
each know how to build and manage selectors and payloads.

This also refactors SelectorEvaluator to build an intermediate stage,
effectively currying until the final evaluation stage within the
instance methods themselves.

This is an intermediate change in preparation for a larger refactor of
all method generation to separate classes.