Adjust RegisteredNameVisitor to better handle name overlaps
What?
=====

This fixes buggy behavior within RegisteredNameVisitor where dynamic
elements with methods that contained nested macros were raising
DuplicateElementDeclarationErrors.

The previous logic to track name registration only used an array, which
didn't take into account that different layers of macros may want to
use the same name.

Now that the construction strategy is passed, we can bind to the object
id of the parent to track names. In the case where a has_one is declared
dynamic selector and the method definition occurs further along in the
file, the strategy gets rebuilt (resulting in a new object_id),
effectively clearing the cache of already-registered elements correctly.