Add additional tests to showcase other use-cases (#33)
This commit adds a test to verify that a selector with dynamic options
can use keyword arguments to generate dynamic options.

Additionally, an additional method and expectation is defined to show
how to define predicate matchers for a parent element that utilize a
child element for assertion. I couldn't find an example of this pattern
already in the tests, but if you feel this is already covered please let
me know and I'll update the commit.