Track name registration for delegated methods
What?
=====

In addition to macro declaration and defining methods normally,
developers can delegate methods to other instance methods.

This decorates `delegate` to track the "final form" of each method (as
the `to` and `prefix` options impact the resulting method name) and
handle collisions as it does in other instances. Just as in other
instances, the depth of nesting impacts collisions appropriately.